fix: replace BigInt64Array with Uint32Array for perceptual hashing
BigInt operations are ~50x slower than regular number arithmetic in
browser JS engines, causing the matcher to hang for minutes on even
small frame counts. Switching to 32-bit dHash stored in Uint32Array
with popcount-based Hamming distance brings matching from 'hung' to
<30ms for typical workloads.

- computeDHash now produces a 32-bit hash (4 rows × 8 bits)
- hashSimilarity uses SWAR popcount on regular numbers
- FingerprintStream.hashes type changed to Uint32Array | null
- Worker and matcher updated accordingly

+++ src/lib/comparison/types.ts
@@ -46,8 +46,8 @@ export interface FingerprintStream {
   times: Float32Array;
   /** Feature vectors — interpretation depends on `kind`. */
   vectors: Float32Array[];
-  /** For hash streams: packed 64-bit hashes as BigInt64Array (else null). */
-  hashes: BigInt64Array | null;
+  /** For hash streams: packed 32-bit hashes as Uint32Array (else null). */
+  hashes: Uint32Array | null;
   /** Original length of the source media this stream was derived from. */
   sourceDuration: number;
 }
